+++ spec/parsed_section_spec.rb
@@ -1,13 +1,12 @@
-require 'paxmex/parsed_section'
+require 'spec_helper'
 
 describe Paxmex::ParsedSection do
   let(:section) { Paxmex::Schema::Section.new('BABA', {}) }
   subject(:parsed_section) { Paxmex::ParsedSection.new(section, {}) }
 
-  it { should be_a_kind_of(Hash) }
+  it { is_expected.to be_a_kind_of(Hash) }
 
   it 'has a key' do
-    parsed_section.key.should == 'BABA'
+    expect(parsed_section.key).to eq('BABA')
   end
-
 end