+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.0.0
+* Allow loading content from a string as well as from a file.
+* Introduce new public interface for parsing and loading, eg. Paxmex.epraw.parse(...).
+* Upgrade to rspec 3.0.
+
 # 1.2.1
 * General cleanup
 * Fixed issue with numeric fields containing float values.